tests/test_cli.py
from click.testing import CliRunner
from webtomd.cli import app


def test_version_flag():
    runner = CliRunner()
    result = runner.invoke(app, ["--version"])
    assert result.exit_code == 0
    assert result.output.strip()

